Add: Age Check, add: Age Validity, add: Check odd even, add: Shopping Cart, add: String Equality.
Age check that return depends on the age, the person can drive.
Age validity that returns a boolean value.
Check odd or even, return if the input value by the user is an even or and odd number.
Shopping cart returns the total price and change the message if the quantity of items is more than 1 to a plural message.
String equality returns if the input name by the user is equals to moe, prints a message else prints another message.